#19 Use RecurringType from PaymentProvider also at a second position  - and also remove obsolete constant
